Bugfix: Transactions signing
The previous transaction signing system used SIGN_MODE_AMINO_JSON, which is deprecated and no longer supported on the SEKAI side. This branch introduces a new way to sign transactions that use SIGN_MODE_DIRECT mode.

List of changes:
- imported "cryptography_utils" package which contains crypto-algorithms required to build and sign Cosmos transactions
- removed classes from "infra/dto/api_kira/broadcast/request/transaction/components" as they are included in the “cryptography_utils” package and are no longer needed on the miro side
- modified classes from "infra/dto/shared/messages" to extend "ProtobufAny" class from "cryptography_utils" package and define protobuf serialization for Cosmos messages
- replaced Uint8List private key in wallet.dart with ECPrivateKey from "cryptography_utils" package to simplify accessing private key for transaction signing
- removed signature_utils.dart and sign() method from tx_utils.dart, as now signing functionality is handled by the external package
- created new sign() method in unsigned_tx_model.dart, which calls proper signing method from "cryptography_utils" package
- removed Wallet.fromKeyfileData() constructor from wallet.dart as it wasn't used anywhere but needed improvement after changes
- adjusted tests and classes where updated classes were used

+++ lib/shared/models/transactions/form_models/ir_msg_request_verification_form_model.dart
@@ -43,8 +43,8 @@ class IRMsgRequestVerificationFormModel extends AMsgFormModel {
       walletAddress: _requesterWalletAddress!,
       verifierWalletAddress: _verifierWalletAddress!,
       tipTokenAmountModel: _tipTokenAmountModel!,
-      recordIds: <BigInt>[
-        BigInt.parse(_irRecordModel.id),
+      recordIds: <int>[
+        int.parse(_irRecordModel.id),
       ],
     );
   }
